rename pkg util to helpers
