+++ helpers/testdata/empty_table.golden
