+++ helpers/testdata/multiple_rows.golden
