+++ helpers/testdata/no_options.golden
