+++ helpers/testdata/with_footer_option.golden
