+++ helpers/testdata/with_header_option.golden
