+++ helpers/testdata/with_multiple_options.golden
