+++ helpers/testdata/with_single_option.golden
